+++ src/main/java/com/raoulvdberge/refinedstorage/integration/jei/RecipeTransferHandlerGrid.java
@@ -21,6 +21,9 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 public class RecipeTransferHandlerGrid implements IRecipeTransferHandler {
+    public static final long TRANSFER_SCROLL_DELAY_MS = 200;
+    public static long LAST_TRANSFER;
+
     @Override
     public Class<? extends Container> getContainerClass() {
         return ContainerGrid.class;
@@ -28,6 +31,8 @@ public class RecipeTransferHandlerGrid implements IRecipeTransferHandler {
 
     @Override
     public IRecipeTransferError transferRecipe(Container container, IRecipeLayout recipeLayout, EntityPlayer player, boolean maxTransfer, boolean doTransfer) {
+        LAST_TRANSFER = System.currentTimeMillis();
+
         if (doTransfer) {
             IGrid grid = ((ContainerGrid) container).getGrid();